fix: 初始暂存
